collecting features in CUDControllerTrait
